Add the remaining SI base units of electric current and luminous intensity.
The only remaining SI base unit is the mole which I currently do not think we
need to explicitly model as its own dimension as in most cases this can just be
a scaling factor. Although we probably do want a constant somewhere. I'm going
to hold off on adding it to think about this some more.