Consistently use the plural form for units.
The common case is probably for people to use the units with non-one numbers
(`grams(25)`) or in unit conversions (`power.in(watts)`) where the plural form
is more natural. It does come at the cost of making derived units a bit weirder
(`meters.per(seconds)`) as well as when using quanties of one.