Reduce memory pressure of quantities.
We haven't actually seen any problems in the wild yet but some users have
voiced concerns (#2). This moves all common functions of `Quantity` to a
prototype, leaving quantities with the only own properties of amount, unit and
dimension.

This doesn't actually have any impact on runtime performance (as measured by
the microbenchmarks) but likely does reduce memory footprint of quantities a
bit (not measured).

We are only doing this for `Quantity` and not for `Unit` right now since the
latter is callable, making it somewhat more tricky. It‘s also likely less
important for `Unit` as they are likely going to be constructed less often.