Factor out methods of `Unit` into a prototype.
This decreases the memory footprint of `Unit` by moving shared methods into the
prototype. We haven't actually seen issues with the memory footprint in the
wild but there has been a user concern around it (#2). The change has no
visible impact on runtime performance.

Because units are callable (for succint syntax like `meters(5)`) we need to
employ a bit of a hack where we manually set the prototype for a newly
constructed function object.